feat: Created the 'configs' folder and 'config.ts' file
Imported '*' as 'dotenv' from the 'dotenv' library and invoked the 'config' function from 'dotenv'. Created the 'ENVIRONMENT' constant and assigned it the value of the 'NODE_ENV' environment variable as 'environments', where 'environments' was imported from 'types.d.ts'. Additionally, created the exportable constant 'config' of type 'Config', imported from 'types.d.ts'. This object has properties: 'env' with the value 'ENVIRONMENT', 'port' with the value of the 'PORT' environment variable transformed to a number using parseInt, defaulting to '3000' if the environment variable is not present, and 'isPro' evaluating 'ENVIRONMENT' for 'production'. The 'config.ts' file was created within the 'configs' folder